Changed file disposition
