+++ src/game/common_components.rs
